backend/pyproject.toml
[project]
name = "vendocasa-backend"
version = "0.1.0"
description = "Italian Real Estate Valuation Tool - Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0.36",
    "geoalchemy2>=0.18",
    "psycopg[binary,pool]>=3.2",
    "alembic>=1.14",
    "shapely>=2.0",
    "geopy>=2.4",
    "pandas>=2.2",
    "pydantic-settings>=2.0",
    "lxml>=5.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"
